ai-service/requirements.txt
annotated-doc==0.0.4
annotated-types==0.7.0
anyio==4.12.1
attrs==25.4.0
certifi==2026.1.4
cffi==2.0.0
charset-normalizer==3.4.4
click==8.3.1
colorama==0.4.6
cryptography==46.0.3
distro==1.9.0
fastapi==0.128.0
greenlet==3.3.1
griffe==1.15.0
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
httpx-sse==0.4.3
idna==3.11
jiter==0.12.0
jsonschema==4.26.0
jsonschema-specifications==2025.9.1
mcp==1.25.0
openai==2.15.0
openai-agents==0.7.0
pycparser==3.0
pydantic==2.12.5
pydantic-settings==2.12.0
pydantic_core==2.41.5
PyJWT==2.10.1
python-dotenv==1.2.1
python-multipart==0.0.21
referencing==0.37.0
requests==2.32.5
rpds-py==0.30.0
sniffio==1.3.1
SQLAlchemy==2.0.46
sse-starlette==3.2.0
starlette==0.50.0
tqdm==4.67.1
types-requests==2.32.4.20260107
typing-inspection==0.4.2
typing_extensions==4.15.0
urllib3==2.6.3
uvicorn==0.40.0
